+++ AK/RefPtr.h
@@ -14,9 +14,9 @@
 #include <AK/Traits.h>
 #include <AK/Types.h>
 #ifdef KERNEL
+#    include <Kernel/API/KResult.h>
 #    include <Kernel/Arch/x86/Processor.h>
 #    include <Kernel/Arch/x86/ScopedCritical.h>
-#    include <Kernel/KResult.h>
 #endif
 
 namespace AK {
