+++ Kernel/FileSystem/ISO9660FileSystem.cpp
@@ -5,7 +5,6 @@
  */
 
 #include "ISO9660FileSystem.h"
-#include "Kernel/FileSystem/BlockBasedFileSystem.h"
 #include <AK/CharacterTypes.h>
 #include <AK/Endian.h>
 #include <AK/HashFunctions.h>
@@ -15,9 +14,9 @@
 #include <AK/StringHash.h>
 #include <AK/StringView.h>
 #include <Kernel/Debug.h>
+#include <Kernel/FileSystem/BlockBasedFileSystem.h>
 #include <Kernel/Forward.h>
 #include <Kernel/KBuffer.h>
-#include <Kernel/KResult.h>
 #include <Kernel/UnixTypes.h>
 #include <Kernel/UserOrKernelBuffer.h>
