+++ Kernel/Net/TCPSocket.h
@@ -10,7 +10,7 @@
 #include <AK/HashMap.h>
 #include <AK/SinglyLinkedList.h>
 #include <AK/WeakPtr.h>
-#include <Kernel/KResult.h>
+#include <Kernel/API/KResult.h>
 #include <Kernel/Locking/MutexProtected.h>
 #include <Kernel/Net/IPv4Socket.h>
